adding server side auth because of whitelist

+++ src/main/java/com/Kaljmarik/WGKal/domain/util/StatsRatio.java
@@ -5,7 +5,7 @@ public record StatsRatio(
         double kdRatio,      // frags / (battles-survived)
         double damageRatio   // avg damage per battle
 ) {
-    public static StatsRatio from(long battles, long wins, long survivedBattles, long frags, long damageDealt) {
+    public static StatsRatio from(int battles, int wins, int survivedBattles, int frags, int damageDealt) {
         double winRatio = battles > 0 ? (wins * 100.0 / battles) : 0.0;
         double kdRatio = (battles - survivedBattles) > 0 ? ((double) frags / (battles - survivedBattles)) : 0.0;
         double damageRatio = battles > 0 ? ((double) damageDealt / battles) : 0.0;
